requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.4
defusedxml==0.8.0rc2
Django==5.0
django-ckeditor-5==0.2.11
django-cors-headers==4.3.1
django-cuser==2017.3.16
django-filter==23.5
django-jazzmin==2.6.0
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-tinymce==3.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-api-tracking==1.8.4
drf-yasg==1.21.7
h11==0.14.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
modeltranslation==0.25
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.1.0
simplejson==3.19.2
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
